pyproject.toml
[tool.poetry]
name = "schedule-vvsu"
version = "0.1.0"
description = "Проект для автоматической синхронизации расписания занятий с Google Calendar"
authors = ["Maxim <m.nikolaev.contact@gmail.com>"]
#readme = "README.md"

# Пакет schedule_vvsu лежит в папке src/schedule_vvsu
packages = [
    { include = "schedule_vvsu", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9,<4"
annotated-types = "0.7.0"
anyio = "4.8.0"
APScheduler = "3.11.0"
attrs = "25.1.0"
beautifulsoup4 = "4.12.3"
cachetools = "5.5.1"
certifi = "2024.12.14"
charset-normalizer = "3.4.1"
click = "8.1.8"
dotenv = "0.9.9"
exceptiongroup = "1.2.2"
fastapi = "0.115.7"
google-api-core = "2.24.1"
google-api-python-client = "2.160.0"
google-auth = "2.38.0"
google-auth-httplib2 = "0.2.0"
google-auth-oauthlib = "1.2.1"
googleapis-common-protos = "1.66.0"
h11 = "0.14.0"
httplib2 = "0.22.0"
idna = "3.10"
markdown-it-py = "3.0.0"
mdurl = "0.1.2"
oauthlib = "3.2.2"
outcome = "1.3.0.post0"
proto-plus = "1.26.0"
protobuf = "5.29.3"
pyasn1 = "0.6.1"
pyasn1-modules = "0.4.1"
pydantic = "2.10.6"
pydantic-settings = "2.8.1"
pydantic-core = "2.27.2"
Pygments = "2.19.1"
pyparsing = "3.2.1"
PySocks = "1.7.1"
python-dotenv = "1.0.1"
pytz = "2024.2"
requests = "2.32.3"
requests-oauthlib = "2.0.0"
rich = "13.9.4"
rsa = "4.9"
selenium = "4.28.1"
shellingham = "1.5.4"
sniffio = "1.3.1"
sortedcontainers = "2.4.0"
soupsieve = "2.6"
starlette = "0.45.3"
trio = "0.28.0"
trio-websocket = "0.11.1"
typer = "0.15.2"
typing_extensions = "4.12.2"
tzlocal = "5.3.1"
uritemplate = "4.1.1"
urllib3 = "2.3.0"
websocket-client = "1.8.0"
wsproto = "1.2.0"
webdriver-manager = "^4.0.2"
sqlalchemy = "^2.0.41"
uvicorn = "^0.34.2"
python-multipart = "^0.0.20"
psycopg2-binary = "^2.9.10"
python-dateutil = "^2.9.0.post0"
gunicorn = "^22.0"
passlib = {version = ">=1.7.4", extras = ["bcrypt"]}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
bcrypt = "<4.3.0"

[tool.poetry.scripts]
vvsu-cli = "schedule_vvsu.cli.main:main"

[tool.poetry.group.dev.dependencies]
alembic = "^1.16.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
